% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WAcolors.R
\name{pnw_palette}
\alias{pnw_palette}
\title{PNW Palette Generator}
\usage{
pnw_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of the color palette. Options are \code{Starfish}, \code{Baker}, \code{Bay},
\code{Winter}, \code{Lake}, \code{Sunset}, \code{Baker2}}

\item{n}{Number of colors in the palette. Palletes include 5-8 colors, which can be used discretely,
or if more are desired, used as a gradient. All color palettes are inspired from author's nature photography
of the region, and checked for color blind safety using \href{https://gka.github.io/palettes/#/9|s|00429d,96ffea,ffffe0|ffffe0,ff005e,93003a|1|1}{Chroma.js Color Palette Helper}.
If omitted, all colors are used.}

\item{type}{Usage of palette as "continuous" or "discrete". Continuous usage interpolates between colors to create
a scale of values. If omitted, function assumes continuous if n > length of palette, and discrete if n < length of palette.}
}
\description{
PNW Palette Generator
}
\examples{
pnw_palette("Winter",n=100,type="continuous")
pnw_palette("Bay",3)
}
