% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PNWColors.R
\docType{data}
\name{pnw_palettes}
\alias{pnw_palettes}
\title{Complete list of palettes}
\format{An object of class \code{list} of length 10.}
\usage{
pnw_palettes
}
\description{
Use \code{\link{pnw_palette}} to construct palettes of desired length.
}
\keyword{datasets}
