% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PNWColors.R
\docType{data}
\name{pnw_palettes}
\alias{pnw_palettes}
\title{Complete list of palettes}
\format{An object of class \code{list} of length 14.}
\usage{
pnw_palettes
}
\description{
Use \code{names(pnw_palettes)} to view list of palette names.
Currently:  "Starfish", "Shuksan",  "Bay",      "Winter",   "Lake",     "Sunset",   "Shuksan2",
"Cascades", "Sailboat", "Moth",     "Spring",   "Mushroom", "Sunset2", and "Anemone"
}
\keyword{datasets}
